backend/requirements.txt
flask
flask-socketio
flask-cors
torch
torchvision
numpy
pillow
eventlet
